GP - WIP Map
Should display the background but doesn't work

+++ GameTest/Gameplay/GameManager.h
@@ -13,8 +13,8 @@ class GameManager
 
 #pragma region VARIABLES
 	float m_SpeedMap = 50.f;
-	int m_HeightMap = 100.f; // To do 
-	int m_WidthMap = 100.f; // To do
+	float m_HeightMap = 100.f; // To do 
+	float m_WidthMap = 100.f; // To do
 
 	Player* m_Player = nullptr;
 	MapManager* m_MapManager = nullptr;
